udp: Add SessionUDP, SessionUDPFactory interfaces.
Allow user to replace the UDP implementation by supplying a
SessionUDPFactory.  A default implementation without any behavioral
changes is provided and used when user does not provide their own
implementation.

The SessionUDPs are pooled, replacing the buffer pooling so that space
for OOB/control data does not need to be dynamically allocated for
each request. The pooling logic is moved from server.go to udp.go,
simplifying the server code a bit.

SessionUDP interface contains a LocalAddr() function, the default
implementation of which returns the address the server is listening on
(no functional change). This allows specialized implementations to
return the address the request was sent to when, e.g., the server is
listening on the IPv6 unspecified address ([::]).

This extensibility is needed to allow transparent proxying, where the
listening socket receives requests originally sent to a different
address/port (see
https://www.kernel.org/doc/Documentation/networking/tproxy.txt for
reference).

(This commit is a rebase/update of the original commit fd36f6f5d68af.)

Co-authored-by: Anton Protopopov <aspsk@isovalent.com>
Signed-off-by: Anton Protopopov <aspsk@isovalent.com>