shared client: Fix global timeout not being respected
This fixes a bug where the shared client did not respect
`c.Client.Timeout`, resulting in DNS requests timing out too soon.

A per Go docs, `Client.Timeout` is supposed to also apply to the write
timeout. This is implemented in the regular client by using
`Client.getTimeoutForRequest` to determine the actual timeout. This
commit adapts the shared client code to do the same.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>