shared_client: Try new id if already in use
Only fail out if non-conflicting request id can not be found in couple of
tries.

This works on the premise that the callers are fine with the request id
being modified at this point. Current use sets a random id just prior to
Exchange call, so this premise is satisfied.

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>